Added version of Old Bird detector that processes input in chunks.
The new version is in the module `old_bird_detector_redux`.

Included unit tests for the `_TransientFinder` class used in the new
detector. Writing the unit tests revealed some bugs in the transient
finding algorithm of the initial, non-chunked version of the Old Bird
detector reimplementation. These bugs are fixed in the new version.

Also modified `test_detector` script to compare clips detected by any
two of the original Bird Detector, the non-chunked reimplementation, and
the chunked reimplementation. The reimplementations produce identical
clips for all test cases, in spite of the differences in their transient
finding algorithms. That is not too surprising, since the differences
concern rare situations.